Fixed navigation problems.
The navigation code has been modified store the facet's state
separately in the facet object itself. The path state is stored
in the navigation object. When the path is changed to view a new
facet, only the path and the state of the new facet will be shown
in the URL, thus keeping the URL short.

This fixes pagination, bookmark and search filter problems as well.

Ticket #1507, 1516, 1517