transifex translation adjustment
Pull the new translations for Spanish (es) and Ukrainian (uk)

Update the LINGUAS file to add comment showing the friendly
name for the language abbreviation.

The make target msg-stats which produces a report about the state
of the translations no longer maintained it's column alignment
due to larger numbers so the formating was tweaked to maintain
column alignment.