Fixed missing optional field.
The optional uid field in user's adder dialog did not appear when
the link is clicked to show the field. This is a regression introduced
in the patch for ticket #1648.

The click handler for the link field has been moved into a new closure
so that the variables point to the correct elements.

Note: the duplicate code in IPA.details_table_section.create() and
IPA.dialog.create() will be addressed separately in ticket #1394.